Conflicts:
	urbackupclient/ClientService.cpp
	urbackupclient/ClientService.h
	urbackupcommon/settingslist.cpp
	urbackupserver/dao/ServerBackupDao.cpp
	urbackupserver/doc/admin_guide.tex
	urbackupserver/fileclient/FileClient.cpp
	urbackupserver/fileclient/FileClientChunked.cpp
	urbackupserver/server_download.cpp
	urbackupserver/server_get.cpp
	urbackupserver/server_get.h
	urbackupserver/server_settings.cpp
	urbackupserver/server_settings.h
	urbackupserver/www/templates.js

+++ urbackupcommon/os_functions_lin.cpp
@@ -38,6 +38,7 @@
 #include <sys/types.h>
 #include <fcntl.h>
 #include <sys/ioctl.h>
+#include <errno.h>
 
 #if defined(__FreeBSD__)
 #define lstat64 lstat
@@ -305,7 +306,7 @@ bool os_directory_exists(const std::wstring &path)
 	return isDirectory(path);
 }
 
-bool os_remove_nonempty_dir(const std::wstring &path, os_symlink_callback_t symlink_callback, void* userdata)
+bool os_remove_nonempty_dir(const std::wstring &path, os_symlink_callback_t symlink_callback, void* userdata, bool delete_root)
 {
 	std::string upath=Server->ConvertToUTF8(path);
 	std::vector<SFile> tmp;
@@ -385,6 +386,13 @@ bool os_remove_nonempty_dir(const std::wstring &path, os_symlink_callback_t syml
 				{
 					symlink_callback(Server->ConvertToUnicode(upath+"/"+(std::string)dirp->d_name), userdata);
 				}
+				else
+				{
+					if(unlink((upath+"/"+(std::string)dirp->d_name).c_str())!=0)
+					{
+						Server->Log("Error deleting symlink \""+upath+"/"+(std::string)dirp->d_name+"\"", LL_ERROR);
+					}
+				}
 			}
 			else
 			{
@@ -399,13 +407,16 @@ bool os_remove_nonempty_dir(const std::wstring &path, os_symlink_callback_t syml
     closedir(dp);
     for(size_t i=0;i<subdirs.size();++i)
     {
-		bool b=os_remove_nonempty_dir(path+L"/"+subdirs[i]);
+		bool b=os_remove_nonempty_dir(path+L"/"+subdirs[i], symlink_callback, userdata);
 		if(!b)
 		    ok=false;
     }
-    if(rmdir(upath.c_str())!=0)
+	if(delete_root)
 	{
-		Server->Log("Error deleting directory \""+upath+"\"", LL_ERROR);
+		if(rmdir(upath.c_str())!=0)
+		{
+			Server->Log("Error deleting directory \""+upath+"\"", LL_ERROR);
+		}
 	}
 	return ok;
 }
@@ -554,4 +565,9 @@ void* os_start_transaction()
 
 bool os_finish_transaction(void* transaction)
 {
+}
+
+int64 os_last_error()
+{
+	return errno;
 }
